backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
boto3==1.35.0
botocore==1.35.0
fastmcp==0.4.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.0
celery==5.4.0
bcrypt==4.0.1
passlib==1.7.4
pyjwt==2.10.0
python-jose[cryptography]==3.3.0
cryptography==44.0.0
pydantic
pydantic-settings
email-validator==2.2.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.52.0
aiofiles==24.1.0
pgvector==0.3.5
numpy==2.1.0
python-dotenv==1.0.1
tenacity==9.0.0
orjson==3.10.12
pytest==8.3.0
pytest-asyncio==0.24.0
